# web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Rempart DNS",
    "version": "1.1",
    "description": "API REST de Rempart. Session (cookie rempart_session + en-tête X-Rempart: 1 sur les requêtes qui modifient) ou jeton porteur rmp_…. Les champs inconnus sont refusés (400)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "Réglages"
        ],
        "summary": "Lire les réglages généraux",
        "description": "Portée : read — Consultation. Tout jeton (sauf backup/sync seuls) et tout rôle.",
        "operationId": "get_api_settings",
        "x-rempart-scope": "read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blocking_enabled": {
                      "type": "boolean"
                    },
                    "blocking_mode": {
                      "type": "string"
                    },
                    "log_mode": {
                      "type": "string"
                    },
                    "retention_days": {
                      "type": "integer"
                    },
                    "client_ids": {
                      "type": "string"
                    },
                    "block_cname_cloaking": {
                      "type": "boolean"
                    },
                    "rebind_protection": {
                      "type": "boolean"
                    },
                    "block_doh_canary": {
                      "type": "boolean"
                    },
                    "suggestions": {
                      "type": "boolean"
                    },
                    "dnssec_validation_off": {
                      "type": "boolean"
                    },
                    "time_zone": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "blocking_enabled": true,
                  "blocking_mode": "zero",
                  "log_mode": "stats",
                  "retention_days": 7,
                  "client_ids": "pseudonymize",
                  "block_cname_cloaking": true,
                  "rebind_protection": true,
                  "block_doh_canary": true,
                  "suggestions": false,
                  "dnssec_validation_off": false,
                  "time_zone": "Europe/Paris"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "Réglages"
        ],
        "summary": "Remplacer les réglages généraux",
        "description": "Objet complet attendu (lisez-le d'abord avec GET). Champ inconnu = 400.\n\nPortée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "put_api_settings",
        "x-rempart-scope": "admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blocking_enabled": {
                    "type": "boolean"
                  },
                  "blocking_mode": {
                    "type": "string"
                  },
                  "log_mode": {
                    "type": "string"
                  },
                  "retention_days": {
                    "type": "integer"
                  },
                  "client_ids": {
                    "type": "string"
                  },
                  "block_cname_cloaking": {
                    "type": "boolean"
                  },
                  "rebind_protection": {
                    "type": "boolean"
                  },
                  "block_doh_canary": {
                    "type": "boolean"
                  },
                  "suggestions": {
                    "type": "boolean"
                  },
                  "dnssec_validation_off": {
                    "type": "boolean"
                  },
                  "time_zone": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "blocking_enabled": true,
                "blocking_mode": "zero",
                "log_mode": "stats",
                "retention_days": 7,
                "client_ids": "pseudonymize",
                "block_cname_cloaking": true,
                "rebind_protection": true,
                "block_doh_canary": true,
                "suggestions": false,
                "dnssec_validation_off": false,
                "time_zone": "Europe/Paris"
              }
            }
          }
        }
      }
    },
...
    "/api/resolvers": {
      "get": {
        "tags": [
          "Réglages"
        ],
        "summary": "Résolveurs en amont et transferts conditionnels",
        "description": "Portée : read — Consultation. Tout jeton (sauf backup/sync seuls) et tout rôle.",
        "operationId": "get_api_resolvers",
        "x-rempart-scope": "read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ]
      },
      "put": {
        "tags": [
          "Réglages"
        ],
        "summary": "Changer la résolution",
        "description": "Profil perso ou entreprise. Upstreams : https:// (DoH), tls:// (DoT), udp://, tcp:// ; « #IP » épingle l'adresse. Appliqué à chaud.\n\nPortée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "put_api_resolvers",
        "x-rempart-scope": "admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profile": {
                    "type": "string"
                  },
                  "upstreams": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "forwards": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "domain": {
                          "type": "string"
                        },
                        "servers": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "bootstrap": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "extra_cas": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "profile": "entreprise",
                "upstreams": [
                  "https://dns.quad9.net/dns-query",
                  "tls://one.one.one.one"
                ],
                "forwards": [
                  {
                    "domain": "corp.example",
                    "servers": [
                      "10.0.0.10",
                      "10.0.0.11"
                    ]
                  }
                ],
                "bootstrap": [
                  "9.9.9.9"
                ],
                "extra_cas": ""
              }
            }
          }
        }
      }
    },
...
    "/api/suggestions/dismiss": {
      "post": {
        "tags": [
          "Ma liste (règles)"
        ],
        "summary": "Écarter une suggestion",
        "description": "Portée : admin — Modification de la configuration. Jeton « admin », rôle opérateur ou administrateur.",
        "operationId": "post_api_suggestions_dismiss",
        "x-rempart-scope": "admin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "domain": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "domain": "telemetry.exemple.com"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton rmp_… créé dans Sécurité → Accès API ou par POST /api/tokens."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rempart_session",
        "description": "Cookie posé par POST /api/login."
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}